requirements.txt
# Core ML/DL frameworks
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
accelerate==0.19.0
transformers==4.29.2

# HuggingFace ecosystem
datasets==2.12.0
huggingface-hub==0.15.1

# Diffusers - must be installed from source for LoRA support
# Run after installing above: pip install git+https://github.com/huggingface/diffusers

# Computer Vision
Pillow==9.5.0

# Web UI
gradio==4.25.0

# Data processing
numpy==1.24.3
pandas==2.0.2

# Testing & Development
pytest==8.2.1
pre-commit==3.7.0

# Misc
tqdm==4.65.0
